requirements.txt
streamlit>=1.20.0
pandas>=1.3
numpy
sentence-transformers
faiss-cpu
plotly
python-dotenv
requests
wikipedia
openai
google-generativeai
tqdm
